Split onTruckPlugin into different packages
