+++ onTruckPlugin/src/main/java/com/github/ontruck/controller/ManualController.java
@@ -1,10 +1,8 @@
-package com.github.ontruck;
+package com.github.ontruck.controller;
 
+import com.github.ontruck.driver.IDriver;
+import com.github.ontruck.network.UDPConnection;
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
-
-import static com.github.ontruck.Driver.BRAKE_OP_CODE;
-import static com.github.ontruck.Driver.POWER_OP_CODE;
-import static com.github.ontruck.Driver.STEER_OP_CODE;
 
 public class ManualController {
 
@@ -17,15 +15,15 @@ public class ManualController {
 	// Process UDP Event
 	public void processEvent(byte type, byte payload, byte stateGroup) {
 		switch(type) {
-			case POWER_OP_CODE:
+			case UDPConnection.POWER_OP_CODE:
 				driver.power(payload);
 				break;
 
-			case STEER_OP_CODE:
+			case UDPConnection.STEER_OP_CODE:
 				driver.steer(payload);
 				break;
 
-			case BRAKE_OP_CODE:
+			case UDPConnection.BRAKE_OP_CODE:
 				driver.brake(payload);
 				break;
 			default:
