+++ onTruckPlugin/src/main/java/com/github/ontruck/controller/plan/PlanExecutor.java
@@ -1,5 +1,6 @@
-package com.github.ontruck;
+package com.github.ontruck.controller.plan;
 
+import com.github.ontruck.driver.IDriver;
 import sun.misc.Lock;
 import sun.reflect.generics.reflectiveObjects.NotImplementedException;
