+++ onTruckPlugin/src/main/java/com/github/ontruck/moped/SensorDataCollector.java
@@ -1,4 +1,4 @@
-package com.github.ontruck;
+package com.github.ontruck.moped;
 
 import com.github.moped.jcan.CAN;
