+++ onTruckPlugin/src/main/java/com/github/ontruck/network/TCPConnection.java
@@ -1,4 +1,4 @@
-package com.github.ontruck;
+package com.github.ontruck.network;
 
 
 import java.io.DataInputStream;
