+++ onTruckPlugin/src/main/java/com/github/ontruck/states/MopedState.java
@@ -1,4 +1,4 @@
-package com.github.ontruck;
+package com.github.ontruck.states;
 
 public enum MopedState {
 	Manual,
